Traslado de metodo Generar Imagen a KokyFrame
